Update cornerstone extension to use shared/root webpack configs
